# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cour API
  version: 0.1.0
  description: |
    REST API for Cour, the seasonal-first anime tracking platform.
    All paths are served under `/api/v1`. Errors use a consistent envelope
    with stable machine-readable codes. Anime metadata is sourced from
    AniList (https://anilist.co).
servers:
  - url: /api/v1

paths:
  /anime:
    get:
      operationId: listAnime
      summary: Browse or search the catalog
      description: >
        With `q`, performs ranked fuzzy search. Without it, browses by
        popularity with optional filters.
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 200 }
        - name: genre
          in: query
          schema: { type: string, maxLength: 50 }
        - name: year
          in: query
          schema: { type: integer, minimum: 1940, maximum: 2100 }
        - name: season
          in: query
          schema: { $ref: "#/components/schemas/Season" }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/AiringStatus" }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 24 }
      responses:
        "200":
          description: Matching anime
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnimeList" }
        default:
          $ref: "#/components/responses/Error"

  /anime/{id}:
    get:
      operationId: getAnime
      summary: Anime detail with episodes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: The anime
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnimeDetail" }
        default:
          $ref: "#/components/responses/Error"

  /seasons/{year}/{season}:
    get:
      operationId: getSeason
      summary: Seasonal chart
      parameters:
        - name: year
          in: path
          required: true
          schema: { type: integer, minimum: 1940, maximum: 2100 }
        - name: season
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Season" }
      responses:
        "200":
          description: All titles airing in the season, most popular first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeasonChart" }
        default:
          $ref: "#/components/responses/Error"

  /schedule:
    get:
      operationId: getSchedule
      summary: Airing schedule for a time window
      description: Defaults to the next 7 days. The window is capped at 14 days.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Episodes airing in the window, soonest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleList" }
        default:
          $ref: "#/components/responses/Error"

components:
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  schemas:
    Season:
      type: string
      enum: [WINTER, SPRING, SUMMER, FALL]

    AiringStatus:
      type: string
      enum: [FINISHED, RELEASING, NOT_YET_RELEASED, CANCELLED, HIATUS]

    Format:
      type: string
      enum: [TV, TV_SHORT, MOVIE, SPECIAL, OVA, ONA, MUSIC]

    AnimeSummary:
      type: object
      required:
        [id, slug, title, title_english, cover_image, cover_color, format,
         status, season, season_year, episodes_count, average_score,
         popularity, genres, next_airing_at, next_airing_episode]
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string, description: Cosmetic URL segment derived from the title }
        title: { type: string, description: Romaji title }
        title_english: { type: string, nullable: true }
        cover_image: { type: string, nullable: true }
        cover_color: { type: string, nullable: true, description: Dominant cover color, hex }
        format:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Format" }]
        status: { $ref: "#/components/schemas/AiringStatus" }
        season:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Season" }]
        season_year: { type: integer, nullable: true }
        episodes_count: { type: integer, nullable: true }
        average_score: { type: integer, nullable: true, description: AniList community mean, 0-100 }
        popularity: { type: integer }
        genres:
          type: array
          items: { type: string }
        next_airing_at: { type: string, format: date-time, nullable: true }
        next_airing_episode: { type: integer, nullable: true }

    AnimeDetail:
      allOf:
        - $ref: "#/components/schemas/AnimeSummary"
        - type: object
          required:
            [anilist_id, title_native, synonyms, description, banner_image,
             duration_min, tags, studios, episodes, synced_at]
          properties:
            anilist_id: { type: integer }
            title_native: { type: string, nullable: true }
            synonyms:
              type: array
              items: { type: string }
            description:
              type: string
              nullable: true
              description: Plain text with newlines; AniList markup stripped.
            banner_image: { type: string, nullable: true }
            duration_min: { type: integer, nullable: true }
            tags:
              type: array
              items: { $ref: "#/components/schemas/Tag" }
            studios:
              type: array
              items: { $ref: "#/components/schemas/Studio" }
            episodes:
              type: array
              items: { $ref: "#/components/schemas/Episode" }
            synced_at: { type: string, format: date-time }

    Tag:
      type: object
      required: [name, rank]
      properties:
        name: { type: string }
        rank: { type: integer }

    Studio:
      type: object
      required: [name, is_main]
      properties:
        name: { type: string }
        is_main: { type: boolean }

    Episode:
      type: object
      required: [number, title, airing_at]
      properties:
        number: { type: integer }
        title: { type: string, nullable: true }
        airing_at: { type: string, format: date-time, nullable: true }

    ScheduleEntry:
      type: object
      required: [anime, episode, airing_at]
      properties:
        anime: { $ref: "#/components/schemas/AnimeSummary" }
        episode: { type: integer }
        airing_at: { type: string, format: date-time }

    AnimeList:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/AnimeSummary" }
        page: { $ref: "#/components/schemas/PageInfo" }

    SeasonChart:
      type: object
      required: [season, year, data]
      properties:
        season: { $ref: "#/components/schemas/Season" }
        year: { type: integer }
        data:
          type: array
          items: { $ref: "#/components/schemas/AnimeSummary" }

    ScheduleList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ScheduleEntry" }

    PageInfo:
      type: object
      required: [page, per_page, has_more]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        has_more: { type: boolean }

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: >
                Stable machine-readable code: bad_request, validation_failed,
                unauthorized, forbidden, not_found, conflict, rate_limited,
                internal_error, service_unavailable.
            message: { type: string }
            details: {}