# backend/internal/httpapi/apigen/api.gen.go
// Package apigen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package apigen

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for AiringStatus.
const (
	CANCELLED      AiringStatus = "CANCELLED"
	FINISHED       AiringStatus = "FINISHED"
	HIATUS         AiringStatus = "HIATUS"
	NOTYETRELEASED AiringStatus = "NOT_YET_RELEASED"
	RELEASING      AiringStatus = "RELEASING"
)

// Valid indicates whether the value is a known member of the AiringStatus enum.
func (e AiringStatus) Valid() bool {
	switch e {
	case CANCELLED:
		return true
	case FINISHED:
		return true
	case HIATUS:
		return true
	case NOTYETRELEASED:
		return true
	case RELEASING:
		return true
	default:
		return false
	}
}

// Defines values for Format.
const (
	MOVIE   Format = "MOVIE"
	MUSIC   Format = "MUSIC"
	ONA     Format = "ONA"
	OVA     Format = "OVA"
	SPECIAL Format = "SPECIAL"
	TV      Format = "TV"
	TVSHORT Format = "TV_SHORT"
)

// Valid indicates whether the value is a known member of the Format enum.
func (e Format) Valid() bool {
	switch e {
	case MOVIE:
		return true
	case MUSIC:
		return true
	case ONA:
		return true
	case OVA:
		return true
	case SPECIAL:
		return true
	case TV:
		return true
	case TVSHORT:
		return true
	default:
		return false
	}
}

// Defines values for Season.
const (
	FALL   Season = "FALL"
	SPRING Season = "SPRING"
	SUMMER Season = "SUMMER"
	WINTER Season = "WINTER"
)

// Valid indicates whether the value is a known member of the Season enum.
func (e Season) Valid() bool {
	switch e {
	case FALL:
		return true
	case SPRING:
		return true
	case SUMMER:
		return true
	case WINTER:
		return true
	default:
		return false
	}
}

// AiringStatus defines model for AiringStatus.
type AiringStatus string

// AnimeDetail defines model for AnimeDetail.
type AnimeDetail struct {
	AnilistId int `json:"anilist_id"`

	// AverageScore AniList community mean
	AverageScore *int    `json:"average_score"`
	BannerImage  *string `json:"banner_image"`

	// CoverColor Dominant cover color
	CoverColor *string `json:"cover_color"`
	CoverImage *string `json:"cover_image"`

	// Description Plain text with newlines; AniList markup stripped.
	Description       *string    `json:"description"`
	DurationMin       *int       `json:"duration_min"`
	Episodes          []Episode  `json:"episodes"`
	EpisodesCount     *int       `json:"episodes_count"`
	Format            *Format    `json:"format"`
	Genres            []string   `json:"genres"`
	Id                int64      `json:"id"`
	NextAiringAt      *time.Time `json:"next_airing_at"`
	NextAiringEpisode *int       `json:"next_airing_episode"`
	Popularity        int        `json:"popularity"`
	Season            *Season    `json:"season"`
	SeasonYear        *int       `json:"season_year"`

	// Slug Cosmetic URL segment derived from the title
	Slug     string       `json:"slug"`
	Status   AiringStatus `json:"status"`
	Studios  []Studio     `json:"studios"`
	SyncedAt time.Time    `json:"synced_at"`
	Synonyms []string     `json:"synonyms"`
	Tags     []Tag        `json:"tags"`

	// Title Romaji title
	Title        string  `json:"title"`
	TitleEnglish *string `json:"title_english"`
	TitleNative  *string `json:"title_native"`
}

// AnimeList defines model for AnimeList.
type AnimeList struct {
	Data []AnimeSummary `json:"data"`
	Page PageInfo       `json:"page"`
}

// AnimeSummary defines model for AnimeSummary.
type AnimeSummary struct {
	// AverageScore AniList community mean
	AverageScore *int `json:"average_score"`

	// CoverColor Dominant cover color
	CoverColor        *string    `json:"cover_color"`
	CoverImage        *string    `json:"cover_image"`
	EpisodesCount     *int       `json:"episodes_count"`
	Format            *Format    `json:"format"`
	Genres            []string   `json:"genres"`
	Id                int64      `json:"id"`
	NextAiringAt      *time.Time `json:"next_airing_at"`
	NextAiringEpisode *int       `json:"next_airing_episode"`
	Popularity        int        `json:"popularity"`
	Season            *Season    `json:"season"`
	SeasonYear        *int       `json:"season_year"`

	// Slug Cosmetic URL segment derived from the title
	Slug   string       `json:"slug"`
	Status AiringStatus `json:"status"`

	// Title Romaji title
	Title        string  `json:"title"`
	TitleEnglish *string `json:"title_english"`
}

// Episode defines model for Episode.
type Episode struct {
	AiringAt *time.Time `json:"airing_at"`
	Number   int        `json:"number"`
	Title    *string    `json:"title"`
}

// ErrorEnvelope defines model for ErrorEnvelope.
type ErrorEnvelope struct {
	Error struct {
		// Code Stable machine-readable code: bad_request, validation_failed, unauthorized, forbidden, not_found, conflict, rate_limited, internal_error, service_unavailable.
		Code    string      `json:"code"`
		Details interface{} `json:"details,omitempty"`
		Message string      `json:"message"`
	} `json:"error"`
}

// Format defines model for Format.
type Format string

// PageInfo defines model for PageInfo.
type PageInfo struct {
	HasMore bool `json:"has_more"`
	Page    int  `json:"page"`
	PerPage int  `json:"per_page"`
}

// ScheduleEntry defines model for ScheduleEntry.
type ScheduleEntry struct {
	AiringAt time.Time    `json:"airing_at"`
	Anime    AnimeSummary `json:"anime"`
	Episode  int          `json:"episode"`
}

// ScheduleList defines model for ScheduleList.
type ScheduleList struct {
	Data []ScheduleEntry `json:"data"`
}

// Season defines model for Season.
type Season string

// SeasonChart defines model for SeasonChart.
type SeasonChart struct {
	Data   []AnimeSummary `json:"data"`
	Season Season         `json:"season"`
	Year   int            `json:"year"`
}

// Studio defines model for Studio.
type Studio struct {
	IsMain bool   `json:"is_main"`
	Name   string `json:"name"`
}

// Tag defines model for Tag.
type Tag struct {
	Name string `json:"name"`
	Rank int    `json:"rank"`
}

// Error defines model for Error.
type Error = ErrorEnvelope

// ListAnimeParams defines parameters for ListAnime.
type ListAnimeParams struct {
	Q       *string       `form:"q,omitempty" json:"q,omitempty"`
	Genre   *string       `form:"genre,omitempty" json:"genre,omitempty"`
	Year    *int          `form:"year,omitempty" json:"year,omitempty"`
	Season  *Season       `form:"season,omitempty" json:"season,omitempty"`
	Status  *AiringStatus `form:"status,omitempty" json:"status,omitempty"`
	Page    *int          `form:"page,omitempty" json:"page,omitempty"`
	PerPage *int          `form:"per_page,omitempty" json:"per_page,omitempty"`
}

// GetScheduleParams defines parameters for GetSchedule.
type GetScheduleParams struct {
	From *time.Time `form:"from,omitempty" json:"from,omitempty"`
	To   *time.Time `form:"to,omitempty" json:"to,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Browse or search the catalog
	// (GET /anime)
	ListAnime(w http.ResponseWriter, r *http.Request, params ListAnimeParams)
	// Anime detail with episodes
	// (GET /anime/{id})
	GetAnime(w http.ResponseWriter, r *http.Request, id int64)
	// Airing schedule for a time window
	// (GET /schedule)
	GetSchedule(w http.ResponseWriter, r *http.Request, params GetScheduleParams)
	// Seasonal chart
	// (GET /seasons/{year}/{season})
	GetSeason(w http.ResponseWriter, r *http.Request, year int, season Season)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Browse or search the catalog
// (GET /anime)
func (_ Unimplemented) ListAnime(w http.ResponseWriter, r *http.Request, params ListAnimeParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Anime detail with episodes
// (GET /anime/{id})
func (_ Unimplemented) GetAnime(w http.ResponseWriter, r *http.Request, id int64) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Airing schedule for a time window
// (GET /schedule)
func (_ Unimplemented) GetSchedule(w http.ResponseWriter, r *http.Request, params GetScheduleParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Seasonal chart
// (GET /seasons/{year}/{season})
func (_ Unimplemented) GetSeason(w http.ResponseWriter, r *http.Request, year int, season Season) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListAnime operation middleware
func (siw *ServerInterfaceWrapper) ListAnime(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ListAnimeParams

	// ------------- Optional query parameter "q" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "q", r.URL.Query(), &params.Q, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "q"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "q", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "genre" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "genre", r.URL.Query(), &params.Genre, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "genre"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "genre", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "year" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "year", r.URL.Query(), &params.Year, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "year"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "year", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "season" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "season", r.URL.Query(), &params.Season, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "season"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "season", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "status", r.URL.Query(), &params.Status, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "status"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "status", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page", r.URL.Query(), &params.Page, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "page"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "per_page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "per_page", r.URL.Query(), &params.PerPage, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "per_page"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "per_page", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListAnime(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetAnime operation middleware
func (siw *ServerInterfaceWrapper) GetAnime(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id int64

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAnime(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSchedule operation middleware
func (siw *ServerInterfaceWrapper) GetSchedule(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetScheduleParams

	// ------------- Optional query parameter "from" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "from", r.URL.Query(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "from"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "from", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "to" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "to", r.URL.Query(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "to"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSchedule(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSeason operation middleware
func (siw *ServerInterfaceWrapper) GetSeason(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "year" -------------
	var year int

	err = runtime.BindStyledParameterWithOptions("simple", "year", chi.URLParam(r, "year"), &year, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "year", Err: err})
		return
	}

	// ------------- Path parameter "season" -------------
	var season Season

	err = runtime.BindStyledParameterWithOptions("simple", "season", chi.URLParam(r, "season"), &season, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "season", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSeason(w, r, year, season)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/anime", wrapper.ListAnime)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/anime/{id}", wrapper.GetAnime)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/schedule", wrapper.GetSchedule)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/seasons/{year}/{season}", wrapper.GetSeason)
	})

	return r
}