# backend/internal/store/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"database/sql/driver"
	"fmt"
	"time"
)

type AnimeFormat string

const (
	AnimeFormatTV      AnimeFormat = "TV"
	AnimeFormatTVSHORT AnimeFormat = "TV_SHORT"
	AnimeFormatMOVIE   AnimeFormat = "MOVIE"
	AnimeFormatSPECIAL AnimeFormat = "SPECIAL"
	AnimeFormatOVA     AnimeFormat = "OVA"
	AnimeFormatONA     AnimeFormat = "ONA"
	AnimeFormatMUSIC   AnimeFormat = "MUSIC"
)

func (e *AnimeFormat) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = AnimeFormat(s)
	case string:
		*e = AnimeFormat(s)
	default:
		return fmt.Errorf("unsupported scan type for AnimeFormat: %T", src)
	}
	return nil
}

type NullAnimeFormat struct {
	AnimeFormat AnimeFormat
	Valid       bool // Valid is true if AnimeFormat is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullAnimeFormat) Scan(value interface{}) error {
	if value == nil {
		ns.AnimeFormat, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.AnimeFormat.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullAnimeFormat) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.AnimeFormat), nil
}

type AnimeSeason string

const (
	AnimeSeasonWINTER AnimeSeason = "WINTER"
	AnimeSeasonSPRING AnimeSeason = "SPRING"
	AnimeSeasonSUMMER AnimeSeason = "SUMMER"
	AnimeSeasonFALL   AnimeSeason = "FALL"
)

func (e *AnimeSeason) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = AnimeSeason(s)
	case string:
		*e = AnimeSeason(s)
	default:
		return fmt.Errorf("unsupported scan type for AnimeSeason: %T", src)
	}
	return nil
}

type NullAnimeSeason struct {
	AnimeSeason AnimeSeason
	Valid       bool // Valid is true if AnimeSeason is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullAnimeSeason) Scan(value interface{}) error {
	if value == nil {
		ns.AnimeSeason, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.AnimeSeason.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullAnimeSeason) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.AnimeSeason), nil
}

type AnimeStatus string

const (
	AnimeStatusFINISHED       AnimeStatus = "FINISHED"
	AnimeStatusRELEASING      AnimeStatus = "RELEASING"
	AnimeStatusNOTYETRELEASED AnimeStatus = "NOT_YET_RELEASED"
	AnimeStatusCANCELLED      AnimeStatus = "CANCELLED"
	AnimeStatusHIATUS         AnimeStatus = "HIATUS"
)

func (e *AnimeStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = AnimeStatus(s)
	case string:
		*e = AnimeStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for AnimeStatus: %T", src)
	}
	return nil
}

type NullAnimeStatus struct {
	AnimeStatus AnimeStatus
	Valid       bool // Valid is true if AnimeStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullAnimeStatus) Scan(value interface{}) error {
	if value == nil {
		ns.AnimeStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.AnimeStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullAnimeStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.AnimeStatus), nil
}

type Anime struct {
	ID                int64
	AnilistID         int32
	TitleRomaji       string
	TitleEnglish      *string
	TitleNative       *string
	Synonyms          []string
	Description       *string
	Format            *AnimeFormat
	Status            AnimeStatus
	Season            *AnimeSeason
	SeasonYear        *int32
	EpisodesCount     *int32
	DurationMin       *int32
	Genres            []string
	Tags              []byte
	Studios           []byte
	CoverImage        *string
	CoverColor        *string
	BannerImage       *string
	AverageScore      *int32
	Popularity        int32
	AnilistTrending   int32
	IsAdult           bool
	NextAiringAt      *time.Time
	NextAiringEpisode *int32
	SyncedAt          time.Time
	CreatedAt         time.Time
	UpdatedAt         time.Time
	SearchDoc         interface{}
}

type Episode struct {
	ID        int64
	AnimeID   int64
	Number    int32
	Title     *string
	AiringAt  *time.Time
	CreatedAt time.Time
}